Resolve ID references via helper module
